+++ fusion_report/data/cosmic.py
@@ -16,8 +16,10 @@ class CosmicDB(Db, metaclass=Singleton):
 
     def get_all_fusions(self) -> List[str]:
         """Returns all fusions from database."""
-        query: str = '''SELECT DISTINCT translocation_name
-                        FROM cosmicfusionexport
+        query: str = '''SELECT FIVE_PRIME_GENE_SYMBOL || '--' || THREE_PRIME_GENE_SYMBOL AS fusion_pair
+                            FROM cosmic_fusion_v101_grch38
+                        SELECT DISTINCT fusion_pair
+                        FROM cosmic_fusion_v101_grch38
                         WHERE translocation_name != ""'''
         res = self.select(query)